ENH: Moving classes and methods to helpers.py  (issue #10)
