+++ SliceTrackerUtils/helpers.py
@@ -1,7 +1,89 @@
-import vtk, ctk, qt
+import DICOMLib
+import os, sys
+import slicer, vtk, ctk, qt
 import xml.dom.minidom, datetime
 from Constants import DICOMTAGS
-from mixins import ModuleLogicMixin
+from mixins import ModuleLogicMixin, ModuleWidgetMixin
+
+
+class SmartDICOMReceiver(ModuleLogicMixin):
+
+  def __init__(self, incomingDataDirectory, receiveFinishedCallback):
+    self.incomingDataDirectory = incomingDataDirectory
+    self.receiveFinishedCallback = receiveFinishedCallback
+    self.storeSCPProcess = None
+    self.setupTimers()
+    self.reset()
+
+  def reset(self):
+    self.timerIterations = 0
+    self.startingFileList = []
+    self.currentFileList = []
+    self.dataHasBeenReceived = False
+
+  def setupTimers(self):
+    self.dataReceivedTimer = qt.QTimer()
+    self.dataReceivedTimer.setInterval(5000)
+    self.dataReceivedTimer.timeout.connect(self.checkIfStillSameFileCount)
+    self.dataReceivedTimer.setSingleShot(True)
+
+    self.watchTimer = qt.QTimer()
+    self.watchTimer.setInterval(1000)
+    self.watchTimer.timeout.connect(self.startWatching)
+    self.watchTimer.setSingleShot(True)
+
+  def start(self):
+    self.stop()
+
+    self.startingFileList = self.getFileList(self.incomingDataDirectory)
+    self.lastFileCount = len(self.startingFileList)
+
+    self.storeSCPProcess = DICOMLib.DICOMStoreSCPProcess(incomingDataDir=self.incomingDataDirectory)
+    self.startWatching()
+    self.storeSCPProcess.start()
+    slicer.util.showStatusMessage("Waiting for incoming DICOM data")
+
+  def stop(self):
+    self.stopWatching()
+    if self.storeSCPProcess:
+      self.storeSCPProcess.stop()
+    self.reset()
+
+  def startWatching(self):
+    self.currentFileList = self.getFileList(self.incomingDataDirectory)
+    if self.lastFileCount != len(self.currentFileList):
+      slicer.util.showStatusMessage(self.getReceivingStatusMessage())
+      self.dataHasBeenReceived = True
+      self.lastFileCount = len(self.currentFileList)
+      self.watchTimer.start()
+    elif self.dataHasBeenReceived:
+      self.lastFileCount = len(self.currentFileList)
+      slicer.util.showStatusMessage("DICOM data receive completed.")
+      self.dataReceivedTimer.start()
+    else:
+      self.watchTimer.start()
+
+  def getReceivingStatusMessage(self):
+    if self.timerIterations == 4:
+      self.timerIterations = 0
+    dots = ""
+    for iteration in range(self.timerIterations):
+      dots += "."
+    self.timerIterations += 1
+    return "Receiving DICOM data %s" % dots
+
+  def stopWatching(self):
+    self.dataReceivedTimer.stop()
+    self.watchTimer.stop()
+
+  def checkIfStillSameFileCount(self):
+    self.currentFileList = self.getFileList(self.incomingDataDirectory)
+    if self.lastFileCount == len(self.currentFileList):
+      newFileList = list(set(self.currentFileList) - set(self.startingFileList))
+      self.startingFileList = self.currentFileList
+      self.lastFileCount = len(self.startingFileList)
+      self.receiveFinishedCallback(newFileList=newFileList)
+    self.watchTimer.start()
 
 
 class SliceAnnotation(object):
@@ -229,6 +311,121 @@ class ExtendedQMessageBox(qt.QMessageBox):
     return qt.QMessageBox.exec_(self, *args, **kwargs), self.checkbox.isChecked()
 
 
+class IncomingDataMessageBox(ExtendedQMessageBox):
+
+  def __init__(self, parent=None):
+    super(IncomingDataMessageBox, self).__init__(parent)
+    self.setWindowTitle("Dialog with CheckBox")
+    self.setText("New data has been received. What would you do?")
+    self.setIcon(qt.QMessageBox.Question)
+    trackButton =  self.addButton(qt.QPushButton('Track targets'), qt.QMessageBox.AcceptRole)
+    self.addButton(qt.QPushButton('Postpone'), qt.QMessageBox.NoRole)
+    self.setDefaultButton(trackButton)
+
+
+class RatingWindow(qt.QWidget, ModuleWidgetMixin):
+
+  @property
+  def maximumValue(self):
+    return self._maximumValue
+
+  @maximumValue.setter
+  def maximumValue(self, value):
+    if value < 1:
+      raise ValueError("The maximum rating value cannot be less than 1.")
+    else:
+      self._maximumValue = value
+
+  def __init__(self, maximumValue, text="Please rate the registration result:", *args):
+    qt.QWidget.__init__(self, *args)
+    self.maximumValue = maximumValue
+    self.text = text
+    self.iconPath = os.path.join(os.path.dirname(sys.modules[self.__module__].__file__), 'Resources/Icons')
+    self.setupIcons()
+    self.setLayout(qt.QGridLayout())
+    self.setWindowFlags(qt.Qt.WindowStaysOnTopHint | qt.Qt.FramelessWindowHint)
+    self.setupElements()
+    self.connectButtons()
+    self.showRatingValue = True
+
+  def __del__(self):
+    self.disconnectButtons()
+
+  def isRatingEnabled(self):
+    return not self.disableWidgetCheckbox.checked
+
+  def setupIcons(self):
+    self.filledStarIcon = self.createIcon("icon-star-filled.png", self.iconPath)
+    self.unfilledStarIcon = self.createIcon("icon-star-unfilled.png", self.iconPath)
+
+  def show(self, disableWidget=None, callback=None):
+    self.disabledWidget = disableWidget
+    if disableWidget:
+      disableWidget.enabled = False
+    qt.QWidget.show(self)
+    self.ratingScore = None
+    self.callback = callback
+
+  def setupElements(self):
+    self.layout().addWidget(qt.QLabel(self.text), 0, 0)
+    self.ratingButtonGroup = qt.QButtonGroup()
+    for rateValue in range(1, self.maximumValue+1):
+      attributeName = "button"+str(rateValue)
+      setattr(self, attributeName, self.createButton('', icon=self.unfilledStarIcon))
+      self.ratingButtonGroup.addButton(getattr(self, attributeName), rateValue)
+
+    for button in list(self.ratingButtonGroup.buttons()):
+      button.setCursor(qt.Qt.PointingHandCursor)
+
+    self.ratingLabel = self.createLabel("")
+    row = self.createHLayout(list(self.ratingButtonGroup.buttons()) + [self.ratingLabel])
+    self.layout().addWidget(row, 1, 0)
+
+    self.disableWidgetCheckbox = qt.QCheckBox("Don't display this window again")
+    self.disableWidgetCheckbox.checked = False
+    self.layout().addWidget(self.disableWidgetCheckbox, 2, 0)
+
+  def connectButtons(self):
+    self.ratingButtonGroup.connect('buttonClicked(int)', self.onRatingButtonClicked)
+    for button in list(self.ratingButtonGroup.buttons()):
+      button.installEventFilter(self)
+
+  def disconnectButtons(self):
+    self.ratingButtonGroup.disconnect('buttonClicked(int)', self.onRatingButtonClicked)
+    for button in list(self.ratingButtonGroup.buttons()):
+      button.removeEventFilter(self)
+
+  def eventFilter(self, obj, event):
+    if obj in list(self.ratingButtonGroup.buttons()) and event.type() == qt.QEvent.HoverEnter:
+      self.onHoverEvent(obj)
+    elif obj in list(self.ratingButtonGroup.buttons()) and event.type() == qt.QEvent.HoverLeave:
+      self.onLeaveEvent()
+    return qt.QWidget.eventFilter(self, obj, event)
+
+  def onLeaveEvent(self):
+    for button in list(self.ratingButtonGroup.buttons()):
+      button.icon = self.unfilledStarIcon
+
+  def onHoverEvent(self, obj):
+    ratingValue = 0
+    for button in list(self.ratingButtonGroup.buttons()):
+      button.icon = self.filledStarIcon
+      ratingValue += 1
+      if obj is button:
+        break
+    if self.showRatingValue:
+      self.ratingLabel.setText(str(ratingValue))
+
+  def onRatingButtonClicked(self, buttonId):
+    self.ratingScore = buttonId
+    if self.disabledWidget:
+      self.disabledWidget.enabled = True
+      self.disabledWidget = None
+    if self.callback:
+      self.callback(self.ratingScore)
+    self.hide()
+
+
 class WatchBoxAttribute(object):
 
   @property
@@ -356,20 +553,13 @@ class XMLBasedInformationWatchBox(FileBasedInformationWatchBox):
   def __init__(self, attributes, title="", sourceFile=None, parent=None):
     super(XMLBasedInformationWatchBox, self).__init__(attributes, title, sourceFile, parent)
 
-  def _findElement(self, dom, name):
-    for e in [e for e in dom.getElementsByTagName('element') if e.getAttribute('name') == name]:
-      try:
-        return e.childNodes[0].nodeValue
-      except IndexError:
-        return ""
-
   def updateInformation(self):
     dom = xml.dom.minidom.parse(self._sourceFile)
 
     for attribute in self.attributes:
       values = []
       for tag in attribute.tags:
-        currentValue = self._findElement(dom, tag)
+        currentValue = ModuleLogicMixin.findElement(dom, tag)
         if tag in self.DATE_TAGS_TO_FORMAT:
           currentValue = self.formatDate(currentValue)
         elif tag == "PatientName":
@@ -398,3 +588,125 @@ class DICOMBasedInformationWatchBox(FileBasedInformationWatchBox):
         values.append(currentValue)
       value = self._getTagValueFromTagValues(values)
       self.setInformation(attribute.name, value, toolTip=value)
+
+
+class CustomTargetTableModel(qt.QAbstractTableModel):
+
+  COLUMN_NAME = 'Name'
+  COLUMN_2D_DISTANCE = 'Distance 2D[mm]'
+  COLUMN_3D_DISTANCE = 'Distance 3D[mm]'
+  COLUMN_HOLE = 'Hole'
+  COLUMN_DEPTH = 'Depth [mm]'
+
+  headers = [COLUMN_NAME, COLUMN_2D_DISTANCE, COLUMN_3D_DISTANCE, COLUMN_HOLE, COLUMN_DEPTH]
+
+  @property
+  def targetList(self):
+    return self._targetList
+
+  @targetList.setter
+  def targetList(self, targetList):
+    self.needleStartEndPositions = {}
+    if self._targetList and self.observer:
+      self._targetList.RemoveObserver(self.observer)
+    self._targetList = targetList
+    if self._targetList:
+      self.observer = self._targetList.AddObserver(self._targetList.PointModifiedEvent, self.computeNewDepthAndHole)
+    self.computeNewDepthAndHole()
+    self.reset()
+
+  @property
+  def cursorPosition(self):
+    return self._cursorPosition
+
+  @cursorPosition.setter
+  def cursorPosition(self, cursorPosition):
+    self._cursorPosition = cursorPosition
+    self.dataChanged(self.index(0, 1), self.index(self.rowCount()-1, 2))
+
+  def __init__(self, logic, targets=None, parent=None, *args):
+    qt.QAbstractTableModel.__init__(self, parent, *args)
+    self.logic = logic
+    self._cursorPosition = None
+    self._targetList = None
+    self.needleStartEndPositions = {}
+    self.targetList = targets
+    self.computeCursorDistances = False
+    self.zFrameDepths = {}
+    self.zFrameHole = {}
+    self.observer = None
+    self._targetModifiedCallback = None
+
+  def setTargetModifiedCallback(self, func):
+    assert hasattr(func, '__call__')
+    self._targetModifiedCallback = func
+
+  def headerData(self, col, orientation, role):
+    if orientation == qt.Qt.Horizontal and role in [qt.Qt.DisplayRole, qt.Qt.ToolTipRole]:
+        return self.headers[col]
+    return None
+
+  def rowCount(self):
+    try:
+      number_of_targets = self.targetList.GetNumberOfFiducials()
+      return number_of_targets
+    except AttributeError:
+      return 0
+
+  def columnCount(self):
+    return len(self.headers)
+
+  def data(self, index, role):
+    if not index.isValid() or role not in [qt.Qt.DisplayRole, qt.Qt.ToolTipRole]:
+      return None
+
+    row = index.row()
+    col = index.column()
+
+    targetPosition = [0.0, 0.0, 0.0]
+    if col in [1,2,3,4]:
+      self.targetList.GetNthFiducialPosition(row, targetPosition)
+
+    if col == 0:
+      return self.targetList.GetNthFiducialLabel(row)
+    elif (col == 1 or col == 2) and self.cursorPosition and self.computeCursorDistances:
+      if col == 1:
+        distance2D = self.logic.get2DDistance(targetPosition, self.cursorPosition)
+        return 'x = ' + str(round(distance2D[0], 2)) + ' y = ' + str(round(distance2D[1], 2))
+      distance3D = self.logic.get3DDistance(targetPosition, self.cursorPosition)
+      return str(round(distance3D, 2))
+
+    elif (col == 3 or col == 4) and self.logic.zFrameRegistrationSuccessful:
+      if col == 3:
+        return self.computeZFrameHole(row, targetPosition)
+      else:
+        return self.computeZFrameDepth(row, targetPosition)
+    return ""
+
+  def computeZFrameHole(self, index, targetPosition):
+    if index not in self.zFrameHole.keys():
+      (start, end, indexX, indexY, depth, inRange) = self.logic.computeNearestPath(targetPosition)
+      self.needleStartEndPositions[index] = (start, end)
+      self.zFrameHole[index] = '(%s, %s)' % (indexX, indexY)
+    return self.zFrameHole[index]
+
+  def computeZFrameDepth(self, index, targetPosition):
+    if index not in self.zFrameDepths.keys():
+      (start, end, indexX, indexY, depth, inRange) = self.logic.computeNearestPath(targetPosition)
+      self.zFrameDepths[index] = '%.3f' % depth if inRange else '(%.3f)' % depth
+    return self.zFrameDepths[index]
+
+  def computeNewDepthAndHole(self, observer=None, caller=None):
+    self.zFrameDepths = {}
+    self.zFrameHole = {}
+    if not self.targetList or not self.logic.zFrameRegistrationSuccessful:
+      return
+
+    for index in range(self.targetList.GetNumberOfFiducials()):
+      pos = [0.0, 0.0, 0.0]
+      self.targetList.GetNthFiducialPosition(index, pos)
+      self.computeZFrameHole(index, pos)
+
+    self.dataChanged(self.index(0, 3), self.index(self.rowCount()-1, 4))
+    if self._targetModifiedCallback:
+      self._targetModifiedCallback()
